+++ database/migrations/2020_12_21_091909_create_penjualans_table.php
@@ -16,7 +16,7 @@ class CreatePenjualansTable extends Migration
         Schema::create('penjualans', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('jenis_sampah_id');
-            $table->integer('berat');
+            $table->double('berat');
             $table->timestamps();
         });
     }
